pyproject.toml
[project]
name = "substack-digest"
version = "0.1.0"
description = "Daily AI-powered digest of your Substack subscriptions"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "google-genai>=1.0",
    "resend>=2.0",
    "pydantic-settings>=2.0",
    "jinja2>=3.1",
    "beautifulsoup4>=4.12",
    "yfinance>=1.2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "apscheduler>=3.10,<4",
]

[project.scripts]
substack-digest = "app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
